+++ src/components/BottomNavbar/index.tsx
@@ -14,13 +14,13 @@ const BottomNavbar = () => {
   ];
 
   return (
-    <div className="fixed bottom-0 left-0 w-[100dvw] flex">
+    <div className="fixed bottom-0 left-0 w-[100dvw] flex inset-shadow-sm">
       {routes.map((route) => (
         <div
           key={route.path}
           className={cn(
-            "flex-1 h-[50px] md:h-[50px] flex justify-center items-center cursor-pointer font-normal",
-            `${pathname === route.path ? "bg-blue-100 font-semibold" : "bg-gray-50"}`
+            "text-[14px] md:text-[16px] flex-1 h-[50px] md:h-[50px] flex justify-center items-center cursor-pointer font-normal",
+            `${pathname === route.path ? "font-semibold" : "bg-gray-50"}`
           )}
           onClick={() => {
             navigate(route.path);